+++ packages/msgraph/src/Exceptions/GraphConnectionException.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Moox\Msgraph\Exceptions;
+
+class GraphConnectionException extends GraphException {}